refactor: change Tuple struct types to uint64_t

+++ 01partitioning/partitioning.h
@@ -9,8 +9,8 @@
 #define COMPUTE_HASH_MOD(hash, hash_bit) ((hash) % (hash_bit))
 
 typedef struct {
-    unsigned int key;
-    int value;
+    uint64_t key;
+    uint64_t value;
 } Tuple;
 
 int independent_output(Tuple *tuples, size_t n_tuples, size_t n_hash_bits, size_t n_threads);
